Fix symlink spec on ruby 2.6
The problem with this spec was that it was activating the default
bundler (1.17.2) on a bundle locked to 2.0.1, causing a mismatch error.
We can avoid this by requiring bunlder but not activating its default
gem, and that's precisely what the rubygems `Kernel.require` monkeypatch
takes care of.

Generally we want to avoid this functionality for bundler specs (or
maybe not, now that bundler is a default gem?), but in this case it is
exactly what we want.