Simplify deprecation message
The advice to upgrade the warning doesn't actually work, because there's
an interdependency between this setting and the
`lockfile_uses_separate_rubygems_sources` one, so enabling just one will
print an error.

For now, I'm removing this advice, until we decide what to do with these
two settings. I think the could/should be unified if possible.