Remove `list_command` setting
The list command was still present (just aliased to `bundle show`), even
if the setting was set to false. So the setting was at least super
confusing if not just wrong.

It also led to misleading situations such as

```
$ bundle list --help

(...)

NAME
       bundle-list - List all the gems in the bundle

SYNOPSIS
       bundle list [--name-only] [--paths] [--without-group=GROUP] [--only-group=GROUP]

(...)

$ bundle list --only-group=development
Unknown switches '--only-group=development'
```

So, instead, I enable the new list command _always_ and remove the
`bundle list => bundle show` alias.

As a replacement, I introduce a new `show_command` setting that works
exactly the same as `viz_command` or `console_command`, aiding the
removal of the command but defining the version where it will no longer
work.